% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createmean}
\alias{createmean}
\title{This function calculates the mean}
\usage{
createmean(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
the mean of x
}
\description{
This function calculates the mean
}
\examples{
x <- 1:10
createmean(x)
}

