\name{DDPQuiz3-package}
\alias{DDPQuiz3-package}
\alias{DDPQuiz3}
\docType{package}
\title{
\packageTitle{DDPQuiz3}
}
\description{
\packageDescription{DDPQuiz3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DDPQuiz3}
\packageIndices{DDPQuiz3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DDPQuiz3}

Maintainer: \packageMaintainer{DDPQuiz3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
